renamed field.py to game.py
